yocto: refactor rc-d ruls usage of DISTRO_FEATURES sysvint
The current startup configuration base on rc.d has defects because we use the init manager of busybox.
However, the rc.d related mechanisms of poky are all coupled with DISTRO_FEATURES sysvinit,
which is not conducive to the judgment of the compatibility mechanism of systemd.
In fact, when busybox is used as init manager, its functions are compatible with sysvinit.
For this reason, the following adaptations are made:
    * add DISTRO_FEATURES sysvinit, it can be compatibility with busybox init manager
    * delete old workaround of populate_packages_updatercd
    * init-ifupdown is need for general rc.d startup, fixup its depends
    * fixup rpcbind startup bug: make openeuler patches compatible with our poky project

Signed-off-by: fanglinxu <fanglinxu@huawei.com>