libnsl2-external: avoid the possible conflict of libnsl2
rename the PN of libnsl2-external to avoid the possible conflict of
libnsl2_git.bb where both libnsl2-external and libnsl2_git provide
libnsl2 and libnsl2-dev. Then bitbake will generate a note/warn msg.

libnsl2-external is extracted from external toolchain, and does not
include dev related files, e.g. .h, .a files. So it's not suitable for
build. A better option is to build libnsl2 through libnsl2_git.bb from
the scratch.

In the future, if libnsl2-external can meet the requirement of build.
We can bypass the libnsl2_git to avoid the build of libnsl2.

Signed-off-by: Wayne Ren <renwei41@huawei.com>