openeuler: overrides the do_fetch file-checksum
we can't user poky's original get_checksum_file_list in base.bbclass
because of the src repo organization and special handling of DL_DIR
here we override it with openeuler's implementation. All local files
which start with "file://" and are nt in DL_DIR will be part of checksum
of do_fetch.

So changes of theses files will force do_fetch to run again.

Signed-off-by: Wayne Ren <renwei41@huawei.com>