kernel: downgrade to 5.10.0-153.2.0
For now in openEuler-22.03-LTS-SP3 branch, the supported kernel version
for raspberrypi patch and rt patch are not the same. This will cause
compilation error for raspberrypi. We downgrade kernel version so that
the src-kernel repository contains patches for the same kernel version.

The reason why the do_fetch part of code is deleted is because we use
manifest.yaml as the baseline, changing the commit id is enough.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>