lxcfs: fix mounting error in dmesg
When booting the system up, it outputs error in dmesg continuously:
"new mount options do not match the existing superblock, will be
ignored" every 2.5s.

The reason is, we do not have /var/lib/lxc/lxcfs directory, and
the lxcfs.service keeps mounting this directory through FUSE,
which triggers this error.

Thus, we need to apply a patch to fix this problem. The previous
patch we have in the "files" directory cannot be applied,
so update it.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>