mcs: avoid exporting cpu_logical_map for latest kernel-5.10 version
In the latest kernel-5.10 version, cpu_logical_map is already exported.
If we do the mcs patch, it causes compilation error. However, for boards
like hieulerpi etc., they are still using older version of kernel, in which
the symbol is not exported. Therefore we need to apply the mcs patch for
older versions.

Signed-off-by: Wanming Hu <huwanming@huawei.com>