micrun: fix chmod_modcache to check vendor directory existence
The chmod_modcache function was failing during do_setup_deps prefuncs
stage because it tried to chmod the vendor directory before it was
created. The vendor directory is only created by 'go mod vendor' in
the main do_setup_deps task body.

Split the chmod operations to separately check for the existence of
/pkg and  directories before attempting to change
their permissions. This prevents errors when chmod_modcache is called
before the vendor directory has been created.

Signed-off-by: Shi Xin <shixin21@h-partners.com>